refactor: Extract TiledAxisDebugPlugin to debug.
