+++ examples/helper/anchor.rs
@@ -1,15 +1,6 @@
 use bevy::prelude::*;
 use bevy_ecs_tilemap::prelude::TilemapAnchor;
 
-/// Show the origin with axes.
-pub struct TiledAxisDebugPlugin;
-
-impl Plugin for TiledAxisDebugPlugin {
-    fn build(&self, app: &mut App) {
-        app.add_systems(Update, origin_axes);
-    }
-}
-
 #[allow(dead_code)]
 /// Rotate the tilemap anchor to the right generally but also show custom and
 /// none for completeness.
@@ -28,8 +19,4 @@ pub fn rotate_right(anchor: &TilemapAnchor) -> TilemapAnchor {
         Custom(_) => None,
         None => TopLeft,
     }
-}
-
-fn origin_axes(mut gizmos: Gizmos) {
-    gizmos.axes_2d(Transform::IDENTITY, 1000.0);
 }
